+++ src/audio_super_resolution/__init__.py
@@ -2,8 +2,18 @@
 
 from .audiosr_backend import AUDIOSR_MODEL_NAMES, AUDIOSR_SAMPLE_RATE, AudiosrBackend
 from .config import InferenceConfig, default_model_cache_dir
-from .manifest import build_manifest, write_manifest
+from .manifest import (
+    ManifestComparison,
+    ManifestDifference,
+    build_manifest,
+    compare_manifests,
+    format_manifest_comparison,
+    load_manifest,
+    manifest_comparison_to_dict,
+    write_manifest,
+)
 from .models import ModelInfo, list_models
+from .preprocess import DEFAULT_LOWPASS_CUTOFF_HZ, apply_preprocessing, lowpass_filter
 from .quality import AudioQualityReport, format_quality_report, inspect_audio_quality
 from .resolver import (
     DEFAULT_AUDIO_EXTENSIONS,
@@ -24,19 +34,28 @@ __all__ = [
     "BackendInfo",
     "EnhancementResult",
     "InferenceConfig",
+    "ManifestComparison",
+    "ManifestDifference",
     "ModelInfo",
     "PlannedEnhancement",
     "AudioQualityReport",
     "AUDIOSR_MODEL_NAMES",
     "AUDIOSR_SAMPLE_RATE",
+    "DEFAULT_LOWPASS_CUTOFF_HZ",
+    "apply_preprocessing",
     "available_backends",
     "build_manifest",
+    "compare_manifests",
     "discover_audio_files",
     "default_model_cache_dir",
+    "format_manifest_comparison",
     "format_quality_report",
     "get_backend",
     "inspect_audio_quality",
     "list_models",
+    "lowpass_filter",
+    "load_manifest",
+    "manifest_comparison_to_dict",
     "plan_enhancements",
     "write_manifest",
 ]